Put largest member first in blitdata union
This allows the union to be initialized without narrowing the
initializing expression w.r.t. size, which could result in a bug on
platforms with 16-bit int. (Because union initializations in C++ can
only initialize the first member.) Note that an int initializer must
still be sign-converted to col24, but this shouldn't affect anything.

This also makes the initializing expressions more readable by average,
since most of the time the union is initialized with col24, and not
int. It is initialized with an int only once.

+++ FeLib/Source/feio.cpp
@@ -220,7 +220,7 @@ int iosystem::Menu(cbitmap* BackGround, v2 Pos,
 			    { 0, 0 },
 			    { 0, 0 },
 			    { RES.X, RES.Y },
-			    { static_cast<int>(MakeRGB24(Element, Element, Element)) },
+			    { MakeRGB24(Element, Element, Element) },
 			    0,
 			    0 };
       Backup.LuminanceMaskedBlit(BlitData);
